Create ProductSource implementations
I've created two implementations for retrieving the dress data from the product source, one which uses the John Lewis API directly and one which reads JSON files. Both implementations use the ProductDataConverter to deserialize JSON to the ProductData object.

The deserializer is careful to convert values to empty string of zero if they are null or missing from the product source.